Create appcache file automatically, improve homescreen feature
Toothrot now automatically creates an appcache file when building a game
for browsers. Also, some meta tags were added to the default index.html
file to improve the experience when a user uses the "add to homescreen"
feature on their mobile device.

The appcache file contains a timestamp comment so that the cache is
cleared and rebuild on each new version.