Use dep. injection instead of CommonJS, abstract fs handling
This change means that all engine runtime dependencies are now
build as `build/toothrot.js`.

The change to the file system handling allows the build phase
to happen in the browser using (multiple) node-compatible
in-memory file systems (`fs` module API).

This approach will allow the IDE to be run in the browser
without the need for a backend, including support for plugins
by creating a union of the engine and project file system and
the file systems of the installed plugins during the build
phase for a game.